backend/requirements.txt
# --- Web Framework & API ---
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9

# --- AI & LLM Orchestration ---
langchain-core==0.1.40
langchain-openai==0.1.1
langgraph==0.0.30

# --- Vector Database & Embeddings ---
qdrant-client==1.8.0
sentence-transformers==2.6.1
torch==2.2.1

# --- Primary Database (Local PostgreSQL) ---
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9

# --- Data Processing & Utilities ---
pandas==2.2.1
python-dotenv==1.0.1